Former-commit-id: 8614a60cbb13c8f4640f1073b6c1cc502b39d404

+++ bin/haproxy_get_ssl.py
@@ -1,21 +1,44 @@
 #!/usr/bin/env python  
-import redis, sys, os, json, jinja2
+import redis, sys, os, json, jinja2, pycurl
 from jinja2 import Template
+from StringIO import StringIO
 
 r_server = redis.StrictRedis('127.0.0.1', db=2)
 check = r_server.get("need_CSR")
-check = "1"
-if check == "1":
-    i_key = "owner-info"
-    data=json.loads (r_server.get(i_key))
+if check == None:
+    print "need run registration job" 
+
+if len(sys.argv) > 1:
+    need_haproxy=sys.argv[1]
+else:
+    need_haproxy='no'
+
+#### get hostname
+try:
+    buffer = StringIO()
+    c = pycurl.Curl()
+    c.setopt(c.URL, 'http://169.254.169.253/latest/meta-data/public-hostname')
+    c.setopt(pycurl.CONNECTTIMEOUT, 5)
+    c.setopt(c.WRITEDATA, buffer)
+    c.perform()
+    c.close()
+    hostname = buffer.getvalue()
+except:
+    hostname = "localhost"
+
+
+i_key = "owner-info"
+json_data = r_server.get(i_key)
+if json_data != None:
+    data=json.loads (json_data)
     email = data['Email']
     hostname = data['Hostname']
-    frontend_conf = ""
-    backend_conf = ""
-    acl = ""
-    redirect = ""
+frontend_conf = ""
+backend_conf = ""
+acl = ""
+redirect = ""
 
-
+if check == "1":
 #### get certificate
     if os.environ["ENV"] == 'AWS':
         os.system("mkdir -p /opt/certs/letsencrypt")
@@ -26,9 +49,12 @@ if check == "1":
         os.system("rm -rf /opt/certs/letsencrypt")
         r_server.set("need_CSR", "0")
         r_server.bgsave
-
+else:
+    print "Don't need new certificate"
 
 ### reconfigure haproxy
+if (check == "1") or (need_haproxy == "yes") :
+
     app_key="apps"
     data_apps=r_server.get(app_key)
     os.system("rm -rf /opt/haproxy/haproxy.cfg")
@@ -51,6 +77,16 @@ if check == "1":
 		    acl = acl + ai
 		    ri = "redirect  code 301 location https://kubeapi.%s/api/v1/proxy/namespaces/kube-system/services/monitoring-grafana/ if is_%s \n    " % (data["Hostname"], app["name"])
 		    redirect = redirect + ri
+		elif app["name"] == 'spark':
+		    ai = "acl is_%s hdr_end(host) -i %s.%s \n    " % (app["name"], app["name"], data["Hostname"] )
+		    acl = acl + ai
+		    ri = "redirect  code 301 location https://kubeapi.%s:8080/api/v1/proxy/namespaces/spark-cluster/services/spark-master:8080/ if is_%s \n    " % (data["Hostname"], app["name"])
+		    redirect = redirect + ri
+		elif app["name"] == 'zeppelin':
+		    ai = "acl is_%s hdr_end(host) -i %s.%s \n    " % (app["name"], app["name"], data["Hostname"] )
+		    acl = acl + ai
+		    ri = "redirect  code 301 location https://kubeapi.%s:8080/api/v1/proxy/namespaces/spark-cluster/services/zeppelin/ if is_%s \n    " % (data["Hostname"], app["name"])
+		    redirect = redirect + ri
 		else:
 		    i = "use_backend %s if { hdr_end(host) -i %s }\n    " % (app["name"], app["name"] + "." + data['Hostname']) 
 		    frontend_conf = frontend_conf + i
@@ -66,4 +102,4 @@ if check == "1":
     os.system("haproxy -f /opt/haproxy/haproxy.cfg -p /var/run/haproxy.pid -sf $(cat /var/run/haproxy.pid)")
     
 else:
-    print "Don't need new certificate"
+    print "Don't need reconfigure HAproxy"
